add List component and plugin for hooks, fix App
