+++ src/App.css
@@ -5,14 +5,3 @@
   background-color: #282c34;
   color: #61dafb;
 }
-
-ul{
-  flex-direction: column;
-  border-right: 1px solid #61dafb;
-  padding-right: 2vmin;
-}
-
-li{
-  list-style: none;
-  text-align: left;
-}
