+++ src/App.js
@@ -1,40 +1,14 @@
-import React from 'react';
-import './App.css';
-
-const PATTERNS = [
-  "Abstract Factory",
-  "Builder",
-  "Factory Method",
-  "Prototype",
-  "Singleton",
-  "Adapter",
-  "Bridge",
-  "Composite",
-  "Decorator",
-  "Facade",
-  "Flyweight",
-  "Proxy",
-  "Chain of responsibility",
-  "Command",
-  "Interpreter",
-  "Iterator",
-  "Mediator",
-  "Memento",
-  "Observer",
-  "State",
-  "Strategy",
-  "Template method",
-  "Visitor",
-];
+import React from 'react'
+import List from './components/List/List'
+import './App.css'
 
 function App() {
+
   return (
     <div className="App">
-      <ul>
-        {PATTERNS.map((i, index) => <li key={index}>{i.toString()}</li>)}
-      </ul>
+      <List />
     </div>
-  );
+  )
 }
 
-export default App;
+export default App
